Core updates

+++ app/model/navigation.php
@@ -99,18 +99,16 @@ class App_Model_Navigation extends App_Model_Table
 
 	public function saveGroup($navigation_group_id, $group)
 	{
-		$navigation_group_id = $this->getGroupId($navigation_group_id);
-
-		if (!$navigation_group_id) {
-			return false;
-		}
-
 		if (isset($group['name'])) {
 			if (!validate('text', $group['name'], 3, 64)) {
 				$this->error['name'] = _l("Navigation Group Name must be between 3 and 64 characters!");
 			}
 		} elseif (!$navigation_group_id) {
 			$this->error['name'] = _l("Navigation Group Name is required!");
+		}
+
+		if ($this->error) {
+			return false;
 		}
 
 		if (!isset($group['status'])) {
@@ -131,7 +129,7 @@ class App_Model_Navigation extends App_Model_Table
 		if (!empty($group['links'])) {
 			$this->toTree($group['links']);
 
-			$this->saveGroupLinks($navigation_group_id, $group['links']);
+			$this->saveGroupLinks($navigation_group_id, $group['links'], false);
 		}
 
 		clear_cache('navigation');
@@ -139,7 +137,7 @@ class App_Model_Navigation extends App_Model_Table
 		return $navigation_group_id;
 	}
 
-	public function saveGroupLinks($navigation_group_id, $links)
+	public function saveGroupLinks($navigation_group_id, $links, $append = true)
 	{
 		$navigation_group_id = $this->getGroupId($navigation_group_id);
 
@@ -147,7 +145,9 @@ class App_Model_Navigation extends App_Model_Table
 			return false;
 		}
 
-		$this->delete('navigation', array('navigation_group_id' => $navigation_group_id));
+		if (!$append) {
+			$this->delete('navigation', array('navigation_group_id' => $navigation_group_id));
+		}
 
 		$sort_order = 0;
 
@@ -376,6 +376,10 @@ class App_Model_Navigation extends App_Model_Table
 				$parent_ref[$key] = &$link;
 			}
 
+			if (!isset($link['name'])) {
+				$link['name'] = $key;
+			}
+
 			$parent_ref[$link['name']] = &$link;
 
 			if (!empty($link['parent_id'])) {
@@ -527,11 +531,11 @@ class App_Model_Navigation extends App_Model_Table
 		$this->removeGroup('admin');
 
 		$group = array(
+			'name'   => 'admin',
 			'status' => 1,
-			'stores' => array(-1),
 			'links'  => $links,
 		);
 
-		return $this->saveGroup('admin', $group);
+		return $this->saveGroup(null, $group);
 	}
 }
